+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## Current
 ### Features
 * Added crud filter
+* Added root dashboard for multi campus
 ### Fix
 ## Version 1.4.6
 ### Features